Manage the origin date of the RemindMe request.
The searcher will now save the origin date for the RemindMe request in
the mysql database. The replier will then include that origin date in
the message to the user when their time is up.

Requires an alteration to the database (see the Database scripts file
which has been updated accordingly), but accounts for origin date being
NULL in existing reminders.